Separating menu actions to different classes
