+++ qupath-gui-fx/src/main/java/qupath/lib/gui/actions/ActionTools.java
@@ -19,7 +19,7 @@
  * #L%
  */
 
-package qupath.lib.gui;
+package qupath.lib.gui.actions;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -45,6 +45,7 @@ import org.slf4j.LoggerFactory;
 import javafx.beans.property.Property;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.value.ObservableBooleanValue;
 import javafx.beans.value.ObservableValue;
 import javafx.event.ActionEvent;
 import javafx.scene.Node;
@@ -56,6 +57,9 @@ import javafx.scene.control.ToggleButton;
 import javafx.scene.control.ToggleGroup;
 import javafx.scene.control.Tooltip;
 import javafx.scene.input.KeyCombination;
+import qupath.lib.gui.QuPathGUI;
+import qupath.lib.gui.QuPathResources;
+import qupath.lib.gui.SelectableItem;
 import qupath.lib.gui.tools.IconFactory;
 
 /**
@@ -554,7 +558,7 @@ public class ActionTools {
 	 * @return true if the action has been flagged as selectable, false otherwise.
 	 */
 	public static boolean isSelectable(Action action) {
-		return Boolean.TRUE.equals(action.getProperties().get(qupath.lib.gui.ActionTools.ActionBuilder.Keys.SELECTABLE));
+		return Boolean.TRUE.equals(action.getProperties().get(qupath.lib.gui.actions.ActionTools.ActionBuilder.Keys.SELECTABLE));
 	}
 	
 	/**
@@ -564,7 +568,7 @@ public class ActionTools {
 	 * @see #isSelectable(Action)
 	 */
 	public static void setSelectable(Action action, boolean selectable) {
-		action.getProperties().put(qupath.lib.gui.ActionTools.ActionBuilder.Keys.SELECTABLE, selectable);
+		action.getProperties().put(qupath.lib.gui.actions.ActionTools.ActionBuilder.Keys.SELECTABLE, selectable);
 	}
 	
 	/**
@@ -800,7 +804,7 @@ public class ActionTools {
 				accelerator);
 	}
 	
-	static <T> Action createSelectableCommandAction(final SelectableItem<T> command, final ObservableValue<String> name, final ObservableValue<Node> icon, final KeyCombination accelerator) {
+	public static <T> Action createSelectableCommandAction(final SelectableItem<T> command, final ObservableValue<String> name, final ObservableValue<Node> icon, final KeyCombination accelerator) {
 		var action = ActionTools.actionBuilder(e -> command.setSelected(true))
 				.text(name)
 				.accelerator(accelerator)
@@ -824,6 +828,11 @@ public class ActionTools {
 	 */
 	public static <T> Action createSelectableCommandAction(final SelectableItem<T> command, String name) {
 		return createSelectableCommandAction(command, name, null, null);
+	}
+
+
+	public static Action createSelectableCommandAction(ObservableBooleanValue value) {
+		return createSelectableAction(value, null);
 	}
 
 }
