+++ qupath-gui-fx/src/main/java/qupath/lib/gui/actions/menus/MenuActions.java
@@ -1,0 +1,13 @@
+package qupath.lib.gui.actions.menus;
+
+import java.util.List;
+
+import org.controlsfx.control.action.Action;
+
+public interface MenuActions {
+	
+	List<Action> getActions();
+	
+	String getName();
+	
+}
