+++ qupath-gui-fx/src/main/java/qupath/lib/gui/commands/LogViewerCommand.java
@@ -54,8 +54,8 @@ import javafx.scene.layout.BorderPane;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
 import javafx.stage.Window;
-import qupath.lib.gui.ActionTools;
 import qupath.lib.gui.SelectableItem;
+import qupath.lib.gui.actions.ActionTools;
 import qupath.lib.gui.dialogs.Dialogs;
 import qupath.lib.gui.logging.LogManager;
 import qupath.lib.gui.logging.LogManager.LogLevel;
