+++ qupath-gui-fx/src/main/java/qupath/lib/gui/panes/ProjectBrowser.java
@@ -94,8 +94,8 @@ import javafx.scene.layout.Pane;
 import javafx.scene.layout.StackPane;
 import qupath.lib.common.GeneralTools;
 import qupath.lib.common.ThreadTools;
-import qupath.lib.gui.ActionTools;
 import qupath.lib.gui.QuPathGUI;
+import qupath.lib.gui.actions.ActionTools;
 import qupath.lib.gui.commands.ProjectCommands;
 import qupath.lib.gui.dialogs.Dialogs;
 import qupath.lib.gui.dialogs.Dialogs.DialogButton;
