+++ qupath-gui-fx/src/main/java/qupath/lib/gui/tools/GuiTools.java
@@ -121,8 +121,8 @@ import qupath.lib.color.ColorDeconvolutionStains.DefaultColorDeconvolutionStains
 import qupath.lib.color.StainVector;
 import qupath.lib.common.ColorTools;
 import qupath.lib.common.GeneralTools;
-import qupath.lib.gui.ActionTools;
 import qupath.lib.gui.QuPathGUI;
+import qupath.lib.gui.actions.ActionTools;
 import qupath.lib.gui.commands.Commands;
 import qupath.lib.gui.dialogs.Dialogs;
 import qupath.lib.gui.viewer.QuPathViewer;
