+++ qupath-gui-fx/src/main/resources/qupath/lib/gui/localization/qupath-strings.properties
@@ -211,7 +211,7 @@ Menu.Edit.Copy.description.mainWindowContent = Copy the contents of the main QuP
 Menu.Edit.Copy.description.mainWindowScreenshot = Copy the area of the screen corresponding to the main QuPath window to the clipboard.\nThis includes any additional overlapping windows and dialog boxes.
 Menu.Edit.Copy.description.fullScreenshot = Make a screenshot and copy it to the clipboard.
 Menu.Edit.description.paste = Paste the contents of the system clipboard, if possible.\nIf the clipboard contents are GeoJSON objects, the objects will be pasted to the current image.\nOtherwise, any text found will be shown in a the script editor.
-Menu.Edit.description.pasteToCurrentPlane = aste GeoJSON objects from the system clipboard to the current z-slice and timepoint, if possible.\nNew object IDs will be generated if needed to avoid duplicates.
+Menu.Edit.description.pasteToCurrentPlane = Paste GeoJSON objects from the system clipboard to the current z-slice and timepoint, if possible.\nNew object IDs will be generated if needed to avoid duplicates.
 Menu.Edit.description.resetPreferences = Reset preferences to their default values.\nThis can be useful if you are experiencing any newly-developed persistent problems with QuPath.
 
 # Classify menu
